Module5/my-pipeline/pipeline/assets/ingestion/requirements.txt
pandas==2.2.0
python-dateutil==2.8.2
pyarrow==15.0.0

